pyproject.toml
[project]
name = "auxmet-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "ddgs>=9.5.2",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.116.1",
    "faster-whisper>=1.2.0",
    "groq>=0.30.0",
    "ipykernel>=6.29.5",
    "ipython>=9.4.0",
    "jwt>=1.4.0",
    "kokoro>=0.9.4",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-groq>=0.3.6",
    "langchain-huggingface>=0.3.0",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.6.4",
    "librosa>=0.11.0",
    "pinecone>=7.3.0",
    "pinecone-client>=6.0.0",
    "pinecone-notebooks>=0.1.1",
    "pinecone-text>=0.5.4",
    "pip>=25.2",
    "pydantic>=2.11.7",
    "pydub>=0.25.1",
    "pymongo>=4.13.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "typing-extensions>=4.14.1",
    "uvicorn>=0.35.0",
    "whisper>=1.1.10",
]

